fix build

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -1008,7 +1008,7 @@ void ObjectMgr::CheckCreatureTemplate(CreatureTemplate const* cInfo)
 
     if (cInfo->dmgschool >= SpellSchool(MAX_SPELL_SCHOOL))
     {
-        LOG_ERROR("sql.sql", "Creature (Entry: %u) has invalid spell school value (%u) in `dmgschool`.", cInfo->Entry, cInfo->dmgschool);
+        LOG_ERROR("sql.sql", "Creature (Entry: %u) has invalid spell school value (%u) in `dmgschool`.", cInfo->Entry, uint32(cInfo->dmgschool));
         const_cast<CreatureTemplate*>(cInfo)->dmgschool = SpellSchool::Normal;
     }
